Don't update prompt every 1s - update every 60s
Having the prompt update every 1s significantly interrupts command completion (builtin zsh compinit or zprezto completion module) to annoying extent!
Trade-off:
The prompt is updated every 60s, mainly just to update the git section of the prompt rather than the time.
When a command is executed the time on the command line is then updated to the current time. This has been added to the preexec preexec_start_cmd_timer hook but could be split off later.

Added an option to print the time of command execution starting on the line following the prompt (commented out).